Split url.c into multiple files.
url.c contained four different flavors of external tables:

* "web", using libcurl
* "file", for reading files directly on the server
* "execute", for launching a command and reading its  output.
* "custom", where a custom protocol function does all the work.

It was hard to tell which functions are required by which flavor. To clarify
this, split up url.c into multiple files. What remains in url.c, is just
code to "dispatch" the url_* calls to the correct implementation, depending
on the flavor. All the actual work is done in flavor-specific files.
