+++ gpAux/extensions/gphdfs/gphdfsprotocol.c
@@ -15,6 +15,7 @@
 #include "postgres.h"
 
 #include <fcntl.h>
+#include <unistd.h>
 
 #include "fmgr.h"
 #include "funcapi.h"
@@ -327,7 +328,8 @@ static URL_FILE
 
 	url = quoteArgument(EXTPROTOCOL_GET_URL(fcinfo));
 	initStringInfo(&cmd);
-	appendStringInfo(&cmd, "%s%s %s %s %s", env_cmd.data, java_cmd, format,
+
+	appendStringInfo(&cmd, EXEC_URL_PREFIX "%s%s %s %s %s", env_cmd.data, java_cmd, format,
 			gp_hadoop_connector_version, url);
 
 	if (!forwrite)
@@ -345,7 +347,7 @@ static URL_FILE
 	 * don't have access to the scan counter at all. It's ok because we don't need it.
 	 */
 	external_set_env_vars(&extvar, url, false, NULL, NULL, false, 0);
-	myData = url_execute_fopen(url, cmd.data, forwrite, &extvar);
+	myData = url_execute_fopen(cmd.data, forwrite, &extvar, NULL, NULL, NULL);
 
 	/* Free the command string */
 	pfree(cmd.data);
