+++ src/backend/access/external/Makefile
@@ -9,7 +9,7 @@ subdir = src/backend/access/external
 top_builddir = ../../../..
 include $(top_builddir)/src/Makefile.global
 
-OBJS = fileam.o url.o
+OBJS = fileam.o url.o url_curl.o url_file.o url_execute.o url_custom.o
 
 include $(top_srcdir)/src/backend/common.mk
