+++ src/backend/access/external/test/Makefile
@@ -2,8 +2,8 @@ subdir=src/backend/access/external
 top_builddir=../../../../..
 include $(top_builddir)/src/Makefile.global
 
-TARGETS=url
+TARGETS=url_curl
 
 include $(top_builddir)/src/backend/mock.mk
 
-url.t: $(MOCK_DIR)/backend/cdb/cdbutil_mock.o
+url_curl.t: $(MOCK_DIR)/backend/cdb/cdbutil_mock.o
